Restructured the ChessServer\Command namespace
