+++ src/Command/Play/CommandParser.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChessServer\Command;
+namespace ChessServer\Command\Play;
 
 use ChessServer\Exception\ParserException;
