+++ src/Command/Play/RematchCommand.php
@@ -1,22 +1,28 @@
 <?php
 
-namespace ChessServer\Command;
+namespace ChessServer\Command\Play;
 
 use ChessServer\Game\PlayMode;
 use ChessServer\Socket\ChesslaBlabSocket;
 
-class ResignCommand extends AbstractCommand
+class RematchCommand extends AbstractCommand
 {
     const ACTION_ACCEPT    = 'accept';
 
+    const ACTION_DECLINE   = 'decline';
+
+    const ACTION_PROPOSE   = 'propose';
+
     public function __construct()
     {
-        $this->name = '/resign';
-        $this->description = 'Allows to resign a game.';
+        $this->name = '/rematch';
+        $this->description = 'Allows to offer a rematch.';
         $this->params = [
             // mandatory param
             'action' => [
                 self::ACTION_ACCEPT,
+                self::ACTION_DECLINE,
+                self::ACTION_PROPOSE,
             ],
         ];
     }
