+++ src/Command/Play/UndoCommand.php
@@ -1,25 +1,21 @@
 <?php
 
-namespace ChessServer\Command;
+namespace ChessServer\Command\Play;
 
 use ChessServer\Game\PlayMode;
 use ChessServer\Socket\ChesslaBlabSocket;
 
-class PlayLanCommand extends AbstractCommand
+class UndoCommand extends AbstractCommand
 {
     public function __construct()
     {
-        $this->name = '/play_lan';
-        $this->description = 'Plays a chess move in long algebraic notation.';
-        $this->params = [
-            'color' => '<string>',
-            'lan' => '<string>',
-        ];
+        $this->name = '/undo';
+        $this->description = 'Undoes the last move.';
     }
 
     public function validate(array $argv)
     {
-        return count($argv) - 1 === count($this->params);
+        return count($argv) - 1 === 0;
     }
 
     public function run(ChesslaBlabSocket $socket, array $argv, int $id)
